+++ PinnyNotes.WpfUi/Enums/TransparencyModes.cs
@@ -1,0 +1,8 @@
+﻿namespace PinnyNotes.WpfUi.Enums;
+
+public enum TransparencyModes
+{
+    Disabled,
+    Enabled,
+    WhenPinned
+}