+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 - Updated to .NET 9.
 - General code refactoring.
+- Simplified transparency settings.
 - 
 
 ### Fixed